+++ src/containers/Dropdown/Dropdown.scss
@@ -9,4 +9,9 @@
 select {
 	font-size: 18pt;
 	background: #fff;
+  max-height: 400px;
+}
+
+optgroup {
+  max-height: 400px;
 }
