+++ src/containers/MarketCapCard/MarketCapCard.test.js
