+++ src/reducers/coinReducer.js
@@ -1,7 +1,7 @@
 const coinReducer = (state = [], action) => {
   switch (action.type) {
   case 'SET_COINS' :
-    return [...state, ...action.coins]
+    return action.coins
   default: 
     return state
   }
